Restructure component file structure for dashboard and forms
